build(api): move user controller
